Fix: correction des list pour validiter avec les circulaire

+++ List/libft_list.h
@@ -6,7 +6,7 @@
 /*   By: kebertra <kebertra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/27 18:13:46 by kebertra          #+#    #+#             */
-/*   Updated: 2025/11/28 18:39:59 by kebertra         ###   ########.fr       */
+/*   Updated: 2025/11/29 20:58:11 by kebertra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,5 +32,5 @@ void	ft_lstadd_back(t_list **lst, t_list *new);
 void	ft_lstdelone(t_list *lst, void (*del)(void *));
 void	ft_lstclear(t_list **lst, void (*del)(void *));
 void	ft_lstiter(t_list *lst, void (*f)(void *));
-int		ft_lstsize(t_list *lst);
+size_t	ft_lstsize(t_list *lst);
 #endif
